Simplify code and use one variable to manage which packages to install.
If srvadmin_install_all is set to true install the srvadmin-all which
pulls in the java stuff that is out of date.

If srvadmin_install_all is set to false, install all the packages
except for srvadmin-all, srvadmin-jre, srvadmin-tomcat,
srvadmin-webserver.

And If srvadmin-jre is installed will uninstall it which will also
uninstall srvadmin-all, srvadmin-jre, srvadmin-tomcat,
srvadmin-webserver.